Moved shape implementation classes to their own files
