+++ src/servers/jolt_physics_server_3d.cpp
@@ -9,7 +9,15 @@
 #include "objects/jolt_area_impl_3d.hpp"
 #include "objects/jolt_body_impl_3d.hpp"
 #include "servers/jolt_project_settings.hpp"
-#include "shapes/jolt_shape_impl_3d.hpp"
+#include "shapes/jolt_box_shape_impl_3d.hpp"
+#include "shapes/jolt_capsule_shape_impl_3d.hpp"
+#include "shapes/jolt_concave_polygon_shape_impl_3d.hpp"
+#include "shapes/jolt_convex_polygon_shape_impl_3d.hpp"
+#include "shapes/jolt_cylinder_shape_impl_3d.hpp"
+#include "shapes/jolt_height_map_shape_impl_3d.hpp"
+#include "shapes/jolt_separation_ray_shape_impl_3d.hpp"
+#include "shapes/jolt_sphere_shape_impl_3d.hpp"
+#include "shapes/jolt_world_boundary_shape_impl_3d.hpp"
 #include "spaces/jolt_job_system.hpp"
 #include "spaces/jolt_physics_direct_space_state_3d.hpp"
 #include "spaces/jolt_space_3d.hpp"
