+++ src/shapes/jolt_shape_impl_3d.cpp
@@ -1,7 +1,6 @@
 #include "jolt_shape_impl_3d.hpp"
 
 #include "objects/jolt_object_impl_3d.hpp"
-#include "shapes/jolt_custom_ray_shape.hpp"
 #include "shapes/jolt_custom_user_data_shape.hpp"
 
 JoltShapeImpl3D::~JoltShapeImpl3D() = default;
@@ -162,625 +161,4 @@ void JoltShapeImpl3D::invalidated(bool p_lock) {
 	for (const auto& [owner, ref_count] : ref_counts_by_owner) {
 		owner->shapes_changed(p_lock);
 	}
-}
-
-Variant JoltWorldBoundaryShapeImpl3D::get_data() const {
-	return plane;
-}
-
-void JoltWorldBoundaryShapeImpl3D::set_data(const Variant& p_data) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	ERR_FAIL_COND(p_data.get_type() != Variant::PLANE);
-
-	plane = p_data;
-}
-
-JPH::ShapeRefC JoltWorldBoundaryShapeImpl3D::build() const {
-	ERR_FAIL_D_MSG(
-		"WorldBoundaryShape3D is not supported by Godot Jolt. "
-		"Consider using one or more reasonably sized BoxShape3D instead."
-	);
-}
-
-Variant JoltSeparationRayShapeImpl3D::get_data() const {
-	Dictionary data;
-	data["length"] = length;
-	data["slide_on_slope"] = slide_on_slope;
-	return data;
-}
-
-void JoltSeparationRayShapeImpl3D::set_data(const Variant& p_data) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	ERR_FAIL_COND(p_data.get_type() != Variant::DICTIONARY);
-
-	const Dictionary data = p_data;
-
-	const Variant maybe_length = data.get("length", {});
-	ERR_FAIL_COND(maybe_length.get_type() != Variant::FLOAT);
-
-	const Variant maybe_slide_on_slope = data.get("slide_on_slope", {});
-	ERR_FAIL_COND(maybe_slide_on_slope.get_type() != Variant::BOOL);
-
-	length = maybe_length;
-	slide_on_slope = maybe_slide_on_slope;
-}
-
-String JoltSeparationRayShapeImpl3D::to_string() const {
-	return vformat("{length=%f slide_on_slope=%s}", length, slide_on_slope);
-}
-
-JPH::ShapeRefC JoltSeparationRayShapeImpl3D::build() const {
-	ERR_FAIL_COND_D_MSG(
-		length <= 0.0f,
-		vformat(
-			"Failed to build separation ray shape with %s. "
-			"Its length must be greater than 0.",
-			to_string()
-		)
-	);
-
-	const JoltCustomRayShapeSettings shape_settings(length, slide_on_slope);
-	const JPH::ShapeSettings::ShapeResult shape_result = shape_settings.Create();
-
-	ERR_FAIL_COND_D_MSG(
-		shape_result.HasError(),
-		vformat(
-			"Failed to build separation ray shape with %s. "
-			"It returned the following error: '%s'.",
-			to_string(),
-			to_godot(shape_result.GetError())
-		)
-	);
-
-	return shape_result.Get();
-}
-
-Variant JoltSphereShapeImpl3D::get_data() const {
-	return radius;
-}
-
-void JoltSphereShapeImpl3D::set_data(const Variant& p_data) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	ERR_FAIL_COND(p_data.get_type() != Variant::FLOAT);
-
-	radius = p_data;
-}
-
-String JoltSphereShapeImpl3D::to_string() const {
-	return vformat("{radius=%f}", radius);
-}
-
-JPH::ShapeRefC JoltSphereShapeImpl3D::build() const {
-	ERR_FAIL_COND_D_MSG(
-		radius <= 0.0f,
-		vformat(
-			"Failed to build sphere shape with %s. "
-			"Its radius must be greater than 0.",
-			to_string()
-		)
-	);
-
-	const JPH::SphereShapeSettings shape_settings(radius);
-	const JPH::ShapeSettings::ShapeResult shape_result = shape_settings.Create();
-
-	ERR_FAIL_COND_D_MSG(
-		shape_result.HasError(),
-		vformat(
-			"Failed to build sphere shape with %s. "
-			"It returned the following error: '%s'.",
-			to_string(),
-			to_godot(shape_result.GetError())
-		)
-	);
-
-	return shape_result.Get();
-}
-
-Variant JoltBoxShapeImpl3D::get_data() const {
-	return half_extents;
-}
-
-void JoltBoxShapeImpl3D::set_data(const Variant& p_data) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	ERR_FAIL_COND(p_data.get_type() != Variant::VECTOR3);
-
-	half_extents = p_data;
-}
-
-void JoltBoxShapeImpl3D::set_margin(float p_margin) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	margin = p_margin;
-}
-
-String JoltBoxShapeImpl3D::to_string() const {
-	return vformat("{half_extents=%v margin=%f}", half_extents, margin);
-}
-
-JPH::ShapeRefC JoltBoxShapeImpl3D::build() const {
-	const float shortest_axis = half_extents[half_extents.min_axis_index()];
-
-	ERR_FAIL_COND_D_MSG(
-		shortest_axis <= margin,
-		vformat(
-			"Failed to build box shape with %s. "
-			"Its half extents must be greater than its margin.",
-			to_string()
-		)
-	);
-
-	const JPH::BoxShapeSettings shape_settings(to_jolt(half_extents), margin);
-	const JPH::ShapeSettings::ShapeResult shape_result = shape_settings.Create();
-
-	ERR_FAIL_COND_D_MSG(
-		shape_result.HasError(),
-		vformat(
-			"Failed to build box shape with %s. "
-			"It returned the following error: '%s'.",
-			to_string(),
-			to_godot(shape_result.GetError())
-		)
-	);
-
-	return shape_result.Get();
-}
-
-Variant JoltCapsuleShapeImpl3D::get_data() const {
-	Dictionary data;
-	data["height"] = height;
-	data["radius"] = radius;
-	return data;
-}
-
-void JoltCapsuleShapeImpl3D::set_data(const Variant& p_data) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	ERR_FAIL_COND(p_data.get_type() != Variant::DICTIONARY);
-
-	const Dictionary data = p_data;
-
-	const Variant maybe_height = data.get("height", {});
-	ERR_FAIL_COND(maybe_height.get_type() != Variant::FLOAT);
-
-	const Variant maybe_radius = data.get("radius", {});
-	ERR_FAIL_COND(maybe_radius.get_type() != Variant::FLOAT);
-
-	height = maybe_height;
-	radius = maybe_radius;
-}
-
-String JoltCapsuleShapeImpl3D::to_string() const {
-	return vformat("{height=%f radius=%f}", height, radius);
-}
-
-JPH::ShapeRefC JoltCapsuleShapeImpl3D::build() const {
-	ERR_FAIL_COND_D_MSG(
-		radius <= 0.0f,
-		vformat(
-			"Failed to build capsule shape with %s. "
-			"Its radius must be greater than 0.",
-			to_string()
-		)
-	);
-
-	ERR_FAIL_COND_D_MSG(
-		height <= 0.0f,
-		vformat(
-			"Failed to build capsule shape with %s. "
-			"Its height must be greater than 0.",
-			to_string()
-		)
-	);
-
-	ERR_FAIL_COND_D_MSG(
-		height < radius * 2.0f,
-		vformat(
-			"Failed to build capsule shape with %s. "
-			"Its height must be at least double that of its radius.",
-			to_string()
-		)
-	);
-
-	const float half_height = height / 2.0f;
-	const float cylinder_height = half_height - radius;
-
-	const JPH::CapsuleShapeSettings shape_settings(cylinder_height, radius);
-	const JPH::ShapeSettings::ShapeResult shape_result = shape_settings.Create();
-
-	ERR_FAIL_COND_D_MSG(
-		shape_result.HasError(),
-		vformat(
-			"Failed to build capsule shape with %s. "
-			"It returned the following error: '%s'.",
-			to_string(),
-			to_godot(shape_result.GetError())
-		)
-	);
-
-	return shape_result.Get();
-}
-
-Variant JoltCylinderShapeImpl3D::get_data() const {
-	Dictionary data;
-	data["height"] = height;
-	data["radius"] = radius;
-	return data;
-}
-
-void JoltCylinderShapeImpl3D::set_data(const Variant& p_data) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	ERR_FAIL_COND(p_data.get_type() != Variant::DICTIONARY);
-
-	const Dictionary data = p_data;
-
-	const Variant maybe_height = data.get("height", {});
-	ERR_FAIL_COND(maybe_height.get_type() != Variant::FLOAT);
-
-	const Variant maybe_radius = data.get("radius", {});
-	ERR_FAIL_COND(maybe_radius.get_type() != Variant::FLOAT);
-
-	height = maybe_height;
-	radius = maybe_radius;
-}
-
-void JoltCylinderShapeImpl3D::set_margin(float p_margin) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	margin = p_margin;
-}
-
-String JoltCylinderShapeImpl3D::to_string() const {
-	return vformat("{height=%f radius=%f margin=%f}", height, radius, margin);
-}
-
-JPH::ShapeRefC JoltCylinderShapeImpl3D::build() const {
-	ERR_FAIL_COND_D_MSG(
-		height < margin * 2.0f,
-		vformat(
-			"Failed to build cylinder shape with %s. "
-			"Its height must be at least double that of its margin.",
-			to_string()
-		)
-	);
-
-	ERR_FAIL_COND_D_MSG(
-		radius < margin,
-		vformat(
-			"Failed to build cylinder shape with %s. "
-			"Its radius must be equal to or greater than its margin.",
-			to_string()
-		)
-	);
-
-	const float half_height = height / 2.0f;
-
-	const JPH::CylinderShapeSettings shape_settings(half_height, radius, margin);
-	const JPH::ShapeSettings::ShapeResult shape_result = shape_settings.Create();
-
-	ERR_FAIL_COND_D_MSG(
-		shape_result.HasError(),
-		vformat(
-			"Failed to build cylinder shape with %s. "
-			"It returned the following error: '%s'.",
-			to_string(),
-			to_godot(shape_result.GetError())
-		)
-	);
-
-	return shape_result.Get();
-}
-
-Variant JoltConvexPolygonShapeImpl3D::get_data() const {
-	return vertices;
-}
-
-void JoltConvexPolygonShapeImpl3D::set_data(const Variant& p_data) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	ERR_FAIL_COND(p_data.get_type() != Variant::PACKED_VECTOR3_ARRAY);
-
-	vertices = p_data;
-}
-
-void JoltConvexPolygonShapeImpl3D::set_margin(float p_margin) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	margin = p_margin;
-}
-
-String JoltConvexPolygonShapeImpl3D::to_string() const {
-	return vformat("{vertex_count=%d margin=%f}", vertices.size(), margin);
-}
-
-JPH::ShapeRefC JoltConvexPolygonShapeImpl3D::build() const {
-	const auto vertex_count = (int32_t)vertices.size();
-
-	ERR_FAIL_COND_D_MSG(
-		vertex_count < 3,
-		vformat(
-			"Failed to build convex polygon shape with %s. "
-			"It must have a vertex count of at least 3.",
-			to_string()
-		)
-	);
-
-	JPH::Array<JPH::Vec3> jolt_vertices;
-	jolt_vertices.reserve((size_t)vertex_count);
-
-	const Vector3* vertices_begin = &vertices[0];
-	const Vector3* vertices_end = vertices_begin + vertex_count;
-
-	for (const Vector3* vertex = vertices_begin; vertex != vertices_end; ++vertex) {
-		jolt_vertices.emplace_back(vertex->x, vertex->y, vertex->z);
-	}
-
-	const JPH::ConvexHullShapeSettings shape_settings(jolt_vertices, margin);
-	const JPH::ShapeSettings::ShapeResult shape_result = shape_settings.Create();
-
-	ERR_FAIL_COND_D_MSG(
-		shape_result.HasError(),
-		vformat(
-			"Failed to build convex polygon shape with %s. "
-			"It returned the following error: '%s'.",
-			to_string(),
-			to_godot(shape_result.GetError())
-		)
-	);
-
-	return shape_result.Get();
-}
-
-Variant JoltConcavePolygonShapeImpl3D::get_data() const {
-	Dictionary data;
-	data["faces"] = faces;
-	data["backface_collision"] = backface_collision;
-	return data;
-}
-
-void JoltConcavePolygonShapeImpl3D::set_data(const Variant& p_data) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	ERR_FAIL_COND(p_data.get_type() != Variant::DICTIONARY);
-
-	const Dictionary data = p_data;
-
-	const Variant maybe_faces = data.get("faces", {});
-	ERR_FAIL_COND(maybe_faces.get_type() != Variant::PACKED_VECTOR3_ARRAY);
-
-	const Variant maybe_backface_collision = data.get("backface_collision", {});
-	ERR_FAIL_COND(maybe_backface_collision.get_type() != Variant::BOOL);
-
-	faces = maybe_faces;
-	backface_collision = maybe_backface_collision;
-}
-
-String JoltConcavePolygonShapeImpl3D::to_string() const {
-	return vformat("{vertex_count=%d}", faces.size());
-}
-
-JPH::ShapeRefC JoltConcavePolygonShapeImpl3D::build() const {
-	const auto vertex_count = (int32_t)faces.size();
-	const int32_t face_count = vertex_count / 3;
-	const int32_t excess_vertex_count = vertex_count % 3;
-
-	// HACK(mihe): We can't emit an error for a vertex count of 0 because of things like CSGShape3D
-	// which has its proper initialization deferred, leading to errors being emitted for every
-	// single one that's created
-	QUIET_FAIL_COND_D(vertex_count == 0);
-
-	ERR_FAIL_COND_D_MSG(
-		vertex_count < 3,
-		vformat(
-			"Failed to build concave polygon shape with %s. "
-			"It must have a vertex count of at least 3.",
-			to_string()
-		)
-	);
-
-	ERR_FAIL_COND_D_MSG(
-		excess_vertex_count != 0,
-		vformat(
-			"Failed to build concave polygon shape with %s. "
-			"It must have a vertex count that is divisible by 3.",
-			to_string()
-		)
-	);
-
-	JPH::TriangleList jolt_faces;
-	jolt_faces.reserve((size_t)face_count);
-
-	const Vector3* faces_begin = &faces[0];
-	const Vector3* faces_end = faces_begin + vertex_count;
-
-	for (const Vector3* vertex = faces_begin; vertex != faces_end; vertex += 3) {
-		const Vector3* v0 = vertex + 0;
-		const Vector3* v1 = vertex + 1;
-		const Vector3* v2 = vertex + 2;
-
-		jolt_faces.emplace_back(
-			JPH::Float3(v2->x, v2->y, v2->z),
-			JPH::Float3(v1->x, v1->y, v1->z),
-			JPH::Float3(v0->x, v0->y, v0->z)
-		);
-
-		if (backface_collision) {
-			jolt_faces.emplace_back(
-				JPH::Float3(v0->x, v0->y, v0->z),
-				JPH::Float3(v1->x, v1->y, v1->z),
-				JPH::Float3(v2->x, v2->y, v2->z)
-			);
-		}
-	}
-
-	const JPH::MeshShapeSettings shape_settings(jolt_faces);
-	const JPH::ShapeSettings::ShapeResult shape_result = shape_settings.Create();
-
-	ERR_FAIL_COND_D_MSG(
-		shape_result.HasError(),
-		vformat(
-			"Failed to build concave polygon shape with %s. "
-			"It returned the following error: '%s'.",
-			to_string(),
-			to_godot(shape_result.GetError())
-		)
-	);
-
-	return shape_result.Get();
-}
-
-Variant JoltHeightMapShapeImpl3D::get_data() const {
-	Dictionary data;
-	data["width"] = width;
-	data["depth"] = depth;
-	data["heights"] = heights;
-	return data;
-}
-
-void JoltHeightMapShapeImpl3D::set_data(const Variant& p_data) {
-	ON_SCOPE_EXIT {
-		invalidated();
-	};
-
-	destroy();
-
-	ERR_FAIL_COND(p_data.get_type() != Variant::DICTIONARY);
-
-	const Dictionary data = p_data;
-
-	const Variant maybe_heights = data.get("heights", {});
-	ERR_FAIL_COND(maybe_heights.get_type() != Variant::PACKED_FLOAT32_ARRAY);
-
-	const Variant maybe_width = data.get("width", {});
-	ERR_FAIL_COND(maybe_width.get_type() != Variant::INT);
-
-	const Variant maybe_depth = data.get("depth", {});
-	ERR_FAIL_COND(maybe_depth.get_type() != Variant::INT);
-
-	heights = maybe_heights;
-	width = maybe_width;
-	depth = maybe_depth;
-}
-
-String JoltHeightMapShapeImpl3D::to_string() const {
-	return vformat("{height_count=%d width=%d depth=%d}", heights.size(), width, depth);
-}
-
-JPH::ShapeRefC JoltHeightMapShapeImpl3D::build() const {
-	const auto height_count = (int32_t)heights.size();
-
-	// NOTE(mihe): This somewhat arbitrary limit depends on what the block size is set to, which by
-	// default is 2. If it's set any higher then so would this limit need to be.
-	ERR_FAIL_COND_D_MSG(
-		height_count < 16,
-		vformat(
-			"Failed to build height map shape with %s. "
-			"Height count must be at least 16.",
-			to_string()
-		)
-	);
-
-	ERR_FAIL_COND_D_MSG(
-		height_count != width * depth,
-		vformat(
-			"Failed to build height map shape with %s. "
-			"Height count must be the product of width and depth.",
-			to_string()
-		)
-	);
-
-	ERR_FAIL_COND_D_MSG(
-		width != depth,
-		vformat(
-			"Failed to build height map shape with %s. "
-			"Width must be equal to depth.",
-			to_string()
-		)
-	);
-
-	ERR_FAIL_COND_D_MSG(
-		!is_power_of_2((uint32_t)width),
-		vformat(
-			"Failed to build height map shape with %s. "
-			"Width/depth must be a power-of-two.",
-			to_string()
-		)
-	);
-
-	const int32_t segments_width = width - 1;
-	const int32_t segments_height = depth - 1;
-
-	const float segments_half_width = (float)segments_width / 2.0f;
-	const float segments_half_depth = (float)segments_height / 2.0f;
-
-	const JPH::HeightFieldShapeSettings shape_settings(
-		heights.ptr(),
-		JPH::Vec3(-segments_half_width, 0, -segments_half_depth),
-		JPH::Vec3::sReplicate(1.0f),
-		(JPH::uint32)width
-	);
-
-	// TOOD(mihe): Calculate the necessary bits per sample using `CalculateBitsPerSampleForError`
-
-	const JPH::ShapeSettings::ShapeResult shape_result = shape_settings.Create();
-
-	ERR_FAIL_COND_D_MSG(
-		shape_result.HasError(),
-		vformat(
-			"Failed to build height map shape with %s. "
-			"It returned the following error: '%s'.",
-			to_string(),
-			to_godot(shape_result.GetError())
-		)
-	);
-
-	return shape_result.Get();
 }
