+++ src/shapes/jolt_shape_impl_3d.hpp
@@ -77,218 +77,4 @@ protected:
 	JPH::ShapeRefC jolt_ref;
 };
 
-class JoltWorldBoundaryShapeImpl3D final : public JoltShapeImpl3D {
-public:
-	ShapeType get_type() const override { return ShapeType::SHAPE_WORLD_BOUNDARY; }
-
-	bool is_convex() const override { return false; }
-
-	Variant get_data() const override;
-
-	void set_data(const Variant& p_data) override;
-
-	float get_margin() const override { return 0.0f; }
-
-	void set_margin([[maybe_unused]] float p_margin) override { }
-
-private:
-	JPH::ShapeRefC build() const override;
-
-	Plane plane;
-};
-
-class JoltSeparationRayShapeImpl3D final : public JoltShapeImpl3D {
-public:
-	ShapeType get_type() const override { return ShapeType::SHAPE_SEPARATION_RAY; }
-
-	bool is_convex() const override { return true; }
-
-	Variant get_data() const override;
-
-	void set_data(const Variant& p_data) override;
-
-	float get_margin() const override { return 0.0f; }
-
-	void set_margin([[maybe_unused]] float p_margin) override { }
-
-	String to_string() const;
-
-private:
-	JPH::ShapeRefC build() const override;
-
-	float length = 0.0f;
-
-	bool slide_on_slope = false;
-};
-
-class JoltSphereShapeImpl3D final : public JoltShapeImpl3D {
-public:
-	ShapeType get_type() const override { return ShapeType::SHAPE_SPHERE; }
-
-	bool is_convex() const override { return true; }
-
-	Variant get_data() const override;
-
-	void set_data(const Variant& p_data) override;
-
-	float get_margin() const override { return 0.0f; }
-
-	void set_margin([[maybe_unused]] float p_margin) override { }
-
-	String to_string() const;
-
-private:
-	JPH::ShapeRefC build() const override;
-
-	float radius = 0.0f;
-};
-
-class JoltBoxShapeImpl3D final : public JoltShapeImpl3D {
-public:
-	ShapeType get_type() const override { return ShapeType::SHAPE_BOX; }
-
-	bool is_convex() const override { return true; }
-
-	Variant get_data() const override;
-
-	void set_data(const Variant& p_data) override;
-
-	float get_margin() const override { return margin; }
-
-	void set_margin(float p_margin) override;
-
-	String to_string() const;
-
-private:
-	JPH::ShapeRefC build() const override;
-
-	Vector3 half_extents;
-
-	float margin = 0.04f;
-};
-
-class JoltCapsuleShapeImpl3D final : public JoltShapeImpl3D {
-public:
-	ShapeType get_type() const override { return ShapeType::SHAPE_CAPSULE; }
-
-	bool is_convex() const override { return true; }
-
-	Variant get_data() const override;
-
-	void set_data(const Variant& p_data) override;
-
-	float get_margin() const override { return 0.0f; }
-
-	void set_margin([[maybe_unused]] float p_margin) override { }
-
-	String to_string() const;
-
-private:
-	JPH::ShapeRefC build() const override;
-
-	float height = 0.0f;
-
-	float radius = 0.0f;
-};
-
-class JoltCylinderShapeImpl3D final : public JoltShapeImpl3D {
-public:
-	ShapeType get_type() const override { return ShapeType::SHAPE_CYLINDER; }
-
-	bool is_convex() const override { return true; }
-
-	Variant get_data() const override;
-
-	void set_data(const Variant& p_data) override;
-
-	float get_margin() const override { return margin; }
-
-	void set_margin(float p_margin) override;
-
-	String to_string() const;
-
-private:
-	JPH::ShapeRefC build() const override;
-
-	float height = 0.0f;
-
-	float radius = 0.0f;
-
-	float margin = 0.04f;
-};
-
-class JoltConvexPolygonShapeImpl3D final : public JoltShapeImpl3D {
-public:
-	ShapeType get_type() const override { return ShapeType::SHAPE_CONVEX_POLYGON; }
-
-	bool is_convex() const override { return true; }
-
-	Variant get_data() const override;
-
-	void set_data(const Variant& p_data) override;
-
-	float get_margin() const override { return margin; }
-
-	void set_margin(float p_margin) override;
-
-	String to_string() const;
-
-private:
-	JPH::ShapeRefC build() const override;
-
-	PackedVector3Array vertices;
-
-	float margin = 0.04f;
-};
-
-class JoltConcavePolygonShapeImpl3D final : public JoltShapeImpl3D {
-public:
-	ShapeType get_type() const override { return ShapeType::SHAPE_CONCAVE_POLYGON; }
-
-	bool is_convex() const override { return false; }
-
-	Variant get_data() const override;
-
-	void set_data(const Variant& p_data) override;
-
-	float get_margin() const override { return 0.0f; }
-
-	void set_margin([[maybe_unused]] float p_margin) override { }
-
-	String to_string() const;
-
-private:
-	JPH::ShapeRefC build() const override;
-
-	PackedVector3Array faces;
-
-	bool backface_collision = false;
-};
-
-class JoltHeightMapShapeImpl3D final : public JoltShapeImpl3D {
-public:
-	ShapeType get_type() const override { return ShapeType::SHAPE_HEIGHTMAP; }
-
-	bool is_convex() const override { return false; }
-
-	Variant get_data() const override;
-
-	void set_data(const Variant& p_data) override;
-
-	float get_margin() const override { return 0.0f; }
-
-	void set_margin([[maybe_unused]] float p_margin) override { }
-
-	String to_string() const;
-
-private:
-	JPH::ShapeRefC build() const override;
-
-	PackedFloat32Array heights;
-
-	int32_t width = 0;
-
-	int32_t depth = 0;
-};
-
 #include "jolt_shape_impl_3d.inl"
