+++ src/shapes/jolt_world_boundary_shape_impl_3d.cpp
@@ -1,0 +1,24 @@
+#include "jolt_world_boundary_shape_impl_3d.hpp"
+
+Variant JoltWorldBoundaryShapeImpl3D::get_data() const {
+	return plane;
+}
+
+void JoltWorldBoundaryShapeImpl3D::set_data(const Variant& p_data) {
+	ON_SCOPE_EXIT {
+		invalidated();
+	};
+
+	destroy();
+
+	ERR_FAIL_COND(p_data.get_type() != Variant::PLANE);
+
+	plane = p_data;
+}
+
+JPH::ShapeRefC JoltWorldBoundaryShapeImpl3D::build() const {
+	ERR_FAIL_D_MSG(
+		"WorldBoundaryShape3D is not supported by Godot Jolt. "
+		"Consider using one or more reasonably sized BoxShape3D instead."
+	);
+}
